requirements.txt
mysql-connector-python==9.1.0
simple-term-menu==1.6.6
tabulate==0.9.0
